REGRESSION (276633@main): Block min-content contribution of a scroll container is just the min-content size. https://bugs.webkit.org/show_bug.cgi?id=278645 rdar://135134725
Reviewed by NOBODY (OOPS!).

In 276633@main we changed logicalHeightForGridItem to be more restrictive
when it invalidates a grid item and performs layout on it. This ended
up exposing another bug where we would incorrectly compute the min-content
contribution in the block dimension of a grid item that is also a scroll
container.

In GridTrackSizingAlgorithm::minLogicalSizeForGridItem we compute the
min-content size of a grid item and then end up trying to take into
consideration the min logical height. In this case the min logical height
is auto, but since the grid item is a scroll container it has an
automatic minimum size of 0, so we should just return the min-content
size.

We should only get to this point if the grid item does not have
an automatic minimum size, which is done in minContributionForGridItem
by checking the min logical size property of the item and whether it is
a scroll container, so I added a debug assert also.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-items/overflow-hidden-min-content-contribution-in-abs-pos-grid-in-flexbox-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-items/overflow-hidden-min-content-contribution-in-abs-pos-grid-in-flexbox.html: Added.
* Source/WebCore/rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::GridTrackSizingAlgorithmStrategy::minLogicalSizeForGridItem const):