REGRESSION(281090@main): Sluggish animations on some shopify stores https://bugs.webkit.org/show_bug.cgi?id=279492 rdar://135791322
Reviewed by NOBODY (OOPS!).

In 281090@main, we added an extra bit of information for the grid to keep track of on its
grid items. In particular, we expanded the grid's capabilities to keep track of whether or
not any of its grid items needed to perform stretch alignment. This included some extra
invalidation inside of GridTrackSizingAlgorithm::logicalHeightForGridItem, as performing
layout on a grid item would clear its stretched size, and which is why I still believe this
extra invalidation to be needed.

However, this seems to expose an issue in which we experience thrashing when rerunning the
grid layout. This can become an issue because of 2 main aspects of grid layout currently:

1.) We rerun the full grid track sizing algorithm each time (i.e. little partial layout)
2.) Grid track sizing may need to query the intrinsic sizes of an item. If an item was
previously stretched, we will clear that size, compute the intrinsic sizes for track
sizing, and then stretch the item again.

These 2 issues unfortunately get compounded in the presence of nested grid content.

This patch aims to address 2 by caching the intrinsic logical height of a grid item
to try to avoid the adverse effects of recomputing it. This cache will specifically be
used to cache the intrinsic logical heights of grid items that are computed during the
first pass of row sizing. That means the values that it contains are only valid for that
portion of the track sizing algorithm code, and the track sizing algorithm should only
populate it with sizes during that step.

The cache will live on RenderGrid, which will determine whether or not
it will be available to use by the rest of the code. This availability is dependent on the
type of the current grid content. For now, we will disable this cache if the grid is a
subgrid, if any of the items are a subgrid, if the  grid is a masonry grid, or any of
the items are being baseline aligned (due to the  intrinsic size implications of baseline
aligned grid items in the spec). Depending on the result of this process, we will create
The cache, destroy it, or keep around the existing one.

If we have an existing cache, we will also invalidate the sizes of any grid items which
have been dirtied due to invalidation before performing layout. Also, if the grid
experiences any sort of style change, we currently invalidate the whole cache.

This cache is exposed by RenderGrid via intrinsicLogicalHeightsForRowSizingFirstPass()
so that callers can either query, invalidate, or populate it. Callers should only expect
to be able to use the cache if it is made available by RenderGrid. Inside of logicalHeightForGridItem,
we will check to see if this cache exists and if it contains a value we previously computed
for the item. If we do not have a value for this item, we will just compute it and add
it to the cache afterwards.

We currently invalidate a grid item's cached logical height in minContentContributionForGridItem
and maxContentContributionForGridItem if we end up changing its inline constraints before
we query it for its intrinsic logical height.

* PerformanceTests/Layout/nested-grids-single-item-content-change.html: Added.
* Source/WebCore/rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::GridTrackSizingAlgorithmStrategy::logicalHeightForGridItem const):
(WebCore::GridTrackSizingAlgorithmStrategy::minContentContributionForGridItem const):
(WebCore::GridTrackSizingAlgorithmStrategy::maxContentContributionForGridItem const):
* Source/WebCore/rendering/RenderGrid.cpp:
(WebCore::RenderGrid::styleDidChange):
(WebCore::RenderGrid::layoutGrid):
(WebCore::RenderGrid::isSubgrid const):
(WebCore::RenderGrid::updateIntrinsicLogicalHeightsForRowSizingFirstPassCacheAvailability):
(WebCore::RenderGrid::intrinsicLogicalHeightsForRowSizingFirstPass const):
(WebCore::RenderGrid::canCreateIntrinsicLogicalHeightsForRowSizingFirstPassCache const):
(WebCore::GridItemSizeCache::setSizeForGridItem):
(WebCore::GridItemSizeCache::sizeForItem const):
(WebCore::GridItemSizeCache::invalidateSizeForItem):
* Source/WebCore/rendering/RenderGrid.h: