[GFC] GridLayout should take in UnplacedItems as input https://bugs.webkit.org/show_bug.cgi?id=298729 rdar://160319995
Reviewed by NOBODY (OOPS!).

GridLayout will need some sort of input from GridFormattingContext in
in order to represent the set of items that it will need to run grid
layout on. Since the first step of the grid layout algorithm will be to
perform item placement (https://drafts.csswg.org/css-grid-1/#layout-algorithm),
we can pass in a set of "unplaced items," which will get transformed
into a set of "placed items," as a result of item placement.

In this patch, we introduce the concept of an UnplacedItem, which gets
passed in as part of a list to grid layout. As of right now, this class
is just a simple wrapper around the layout box but in theory will get
populated with enough information about it that allows grid layout to
perform placement. I defer populating the class with that information
until the initial implementation of placement, which is when we will need
it.

Finally, the list which gets sent to grid layout is also logical from
the perspective of the order property. GridFormattingContext sorts this
list based upon the order value so that GridLayout does not have to
worry about it and can just simply iterate over the list without having
to worry about making sure that the order is correct.